lambda2/requirements.txt
boto3
psycopg2-binary
python-dotenv
sqlalchemy